Fix: added some lines to fix but failed

+++ StatisticsFunctions/Z_Score.py
@@ -4,8 +4,8 @@ from StatisticsFunctions.Stand_Dev import standard_deviation
 
 
 def z_score():
-    x_value = mean()
-    y_value = standard_deviation()
+    x_value = float(mean())
+    y_value = float(standard_deviation())
 
     sub = Operations.subtraction(x_value, 5)
     result = Operations.division(y_value, sub)
